requirements.txt
psycopg2-binary==2.9.5
django-axes==5.40.1